Convert favorite services to dropdown.
