+++ src/components/AppFilter/AppFilter.scss
@@ -9,7 +9,7 @@
     bottom:0;
     left: 0;
     top: 70px;
-    overflow-y:scroll;
+    overflow-y:auto;
     overflow-x:hidden;
     position: fixed; // allows menu to go full width
     .chr-app-filter-content {
