--- src/layouts/FavoritedServices.scss
@@ -1,10 +1,0 @@
-@import "~@redhat-cloud-services/frontend-components-utilities/styles/_all";
-@import '~@patternfly/patternfly/patternfly-addons.scss';
-
-.chr-c-favoritedservices {
-  .chr-l-stack__item-centered, .chr-c-card-centered {
-    display: flex;
-    justify-content: center;
-    text-align: center;
-  }
-}
